remove wmts interfaces from ows interfaces
